requirements.txt
# Webアプリケーション
streamlit>=1.28.0

# AI・機械学習
google-generativeai>=0.3.0

# データ処理・分析
pandas>=2.0.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# テスト
pytest>=7.0.0
pytest-cov>=4.0.0 